examples/langgraph/agent/requirements.txt
langgraph>=0.2.0
langchain-openai>=0.1.0
langchain-core>=0.2.0
tavily-python>=0.3.0
